added realtime wallet/service, wallet managment service, wallet info view, icons, and withdraw

+++ lib/core/models/realtime_wallet.dart
@@ -2,7 +2,7 @@ import 'transactions.dart';
 
 class RealTimeWallet {
   final String address;
-  final int balance;
+  final double balance;
   List<Transaction>? transactions;
 
   RealTimeWallet({
